requirements.txt
aniso8601==6.0.0
astroid==2.2.5
autopep8==1.4.4
certifi==2019.3.9
chardet==3.0.4
coverage==4.5.3
coveralls==1.8.1
dj-database-url==0.5.0
Django==2.2.2
django-graphql-jwt==0.2.1
django-heroku==0.3.1
django-nose==1.4.6
docopt==0.6.2
environ==1.0
graphene==2.1.6
graphene-django==2.3.0
graphql-core==2.2
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
nose==1.3.7
promise==2.2.1
psycopg2==2.7.7
psycopg2-binary==2.8.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.3.1
python-decouple==3.1
pytz==2019.1
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
whitenoise==4.1.2
wrapt==1.11.1
